+++ felupe/constitution/__init__.py
@@ -1,14 +1,19 @@
-from ._models import (
+from ._models_hyperelasticity import (
+    NeoHooke,
+)
+
+from ._models_linear_elasticity import (
     LinearElastic,
     LinearElasticTensorNotation,
     LinearElasticPlaneStress,
     LinearElasticPlaneStrain,
-    NeoHooke,
 )
 
-from ._models_tensor import (
+from ._models_pseudo_elasticity import (
     OgdenRoxburgh,
 )
+
+from ._models_umat import UserMaterial
 
 from ._kinematics import (
     LineChange,